Fix unsigned wraparound in instant yield history window computation
UBSan findings 1-4 (CvCultureClasses.cpp:3889/3902):
'2 - 30 cannot be represented in type size_t' and the subsequent
size_t -> int sign-change conversion.

INSTANT_YIELD_HISTORY_LENGTH is declared 'const size_t', so
getGameTurn() (int) is promoted to unsigned before the subtraction.
In the early game (turn < 30) the subtraction wraps; converting the
wrapped value back to int restores the intended negative start turn
on two's complement targets, and getInstantYieldAvg() handles
negative turn bounds correctly, so gameplay behavior was accidentally
correct. The wrap is defined behavior, but the unsigned->int
conversion of an out-of-range value is implementation-defined and a
portability/determinism hazard.

Cast the constant to int so the subtraction happens in signed
arithmetic. No behavior change.

+++ CvGameCoreDLL_Expansion2/CvCultureClasses.cpp
@@ -3886,7 +3886,7 @@ int CvPlayerCulture::GetOtherPlayerCulturePerTurnIncludingInstantTimes100(Player
 {
 	int iBase = GET_PLAYER(eOtherPlayer).GetTotalJONSCulturePerTurnTimes100();
 	int iEndTurn = GC.getGame().getGameTurn();
-	int iStartTurn = GC.getGame().getGameTurn() - INSTANT_YIELD_HISTORY_LENGTH;
+	int iStartTurn = GC.getGame().getGameTurn() - (int)INSTANT_YIELD_HISTORY_LENGTH;
 
 	return iBase + GET_PLAYER(eOtherPlayer).getInstantYieldAvg(YIELD_CULTURE, iStartTurn, iEndTurn) * 100;
 }
@@ -3899,7 +3899,7 @@ int CvPlayerCulture::GetTourismPerTurnIncludingInstantTimes100(PlayerTypes ePlay
 		iBase = GetInfluencePerTurnTimes100(ePlayer);
 
 	int iEndTurn = GC.getGame().getGameTurn();
-	int iStartTurn = GC.getGame().getGameTurn() - INSTANT_YIELD_HISTORY_LENGTH;
+	int iStartTurn = GC.getGame().getGameTurn() - (int)INSTANT_YIELD_HISTORY_LENGTH;
 
 	int iAvgGlobal = m_pPlayer->getInstantYieldAvg(YIELD_TOURISM, iStartTurn, iEndTurn) * 100;
 	int iAvgIndividual = m_pPlayer->getInstantTourismPerPlayerAvg(ePlayer, iStartTurn, iEndTurn) * 100;
